pyproject.toml
[project]
name = "noisemonitor"
version = "1.0.0"
authors = [
  { name="Valérian Fraisse", email="valerian.fraisse@mail.mcgill.ca" },
]
description = "Long-term sound level measurements analysis in Python"
readme = "README.md"
requires-python = ">=3.7"

dependencies = [
    "matplotlib >= 3.5.3",
    "numpy >= 1.21.6",
    "pandas >= 1.3.5",
    "xlrd >= 2.0.1",
    "openpyxl >= 3.1.2",
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
weather = [
    "requests >= 2.25.0",
    "aiohttp >= 3.8.0",
    "lxml >= 4.6.0",
    "python-dateutil >= 2.8.0",
]

[project.urls]
"Homepage" = "https://github.com/valerianF/noisemonitor"
"Bug Tracker" = "https://github.com/valerianF/noisemonitor/issues"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
    "*_test.py",
]
python_classes = [
    "Test*",
]
python_functions = [
    "test_*",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
